+++ CHANGELOG.md
@@ -1,13 +1,27 @@
 Changelog for unzipwalk
 =======================
 
+1.5.0 - Wed, Jun 19 2024
+------------------------
+
+- Added `raise_errors` option that can be turned off so that errors during iteration don't abort the walk,
+  and instead a `UnzipWalkResult` of type `FileType.ERROR` is yielded.
+- **Warning: Potentially incompatible changes**
+  - When `matcher` is used, previously the results would be entirely suppressed, now a `UnzipWalkResult` of type `FileType.SKIP` is yielded.
+  - The CLI tool now defaults to `raise_errors` being off; you must specify `--raise-errors` to get the previous behavior.
+  - The CLI tool now defaults to reporting `ERROR` and `FILE` results instead of just `FILE` results.
+
 1.4.0 - Mon, Jun 17 2024
 ------------------------
+
+`commit c0069e6c6cc20438b70865777b8ca97e8e0dca96`
 
 - Made requirements more lenient to avoid dependency conflicts
 
 1.3.0 - Mon, Jun  3 2024
 ------------------------
+
+`commit 11e5d41703e50348f68410121b078afdcbe30f43`
 
 - Further improved escaping of filenames with `--checksum` CLI option,
   and provided `UnzipWalkResult.from_checksum_line()` to parse the output.